meson: add patch to fix caching of pkg-config files.
https://github.com/void-linux/void-packages/commit/d642a5493d241267a74ce29ab8c4d06fc0fc20db

Allows us to use the PKG_CONFIG variable to tell Meson which is the
pkg-config to be used to get our native dependencies.

We actually could always do that but Meson would cache the variable for
the pkg-config file if we used the dependency() call without 'native:
true' and after that all future usages of dependency() even with
'native: true' would use the cross-compile one.

See this issue from where the patch came from:

- https://github.com/mesonbuild/meson/issues/1736