git: correct ac_cv_fread_reads_directories on musl
git build system relies on FREAD_READS_DIRECTORIES flag to check if
we're on a system which succeeds when attempting to fopen a directory
or read from an fopen-ed directory.

System with musl libc is one of such systems.
The current template lies git build system on this matter.
Effectively, t1308 is failing on musl system.

https://github.com/void-linux/void-packages/commit/bf8bba61b8cde97857bc67be44fec814250be768