build-styles: Add support for Go modules.
Adds support for Go modules by detecting a go.mod file and, if available,
using it. In addition, to continue supporting vendoring (and avoid
the need for git on all module builds), if a package includes a vendor
directory, the module mode will switch to vendor mode if not already
set. This will use vendored source code for dependencies instead of
downloading that code again using the descriptions under go.mod.

go_mod_mode=vendor also skips the go.sum check because nothing is
downloaded that isn't already verified, so this fixes packages with
vendored code that have checksum mismatches due to Go 1.11.4 module
checksum changes.

https://github.com/void-linux/void-packages/commit/43db1c25dcf0247f714fa465067c74ab3b265b39