common: add ppc64 build profiles and other ppc64 bits
This adds the build profiles for ppc64 targets as well as modifications
in other parts of the infra.

These targets are supported:

- ppc64le (glibc little endian elfv2)
- ppc64le-musl (musl little endian)
- ppc64-musl (musl big endian)

ELFv1 targets are explicitly not supported at this point.

Big endian musl supports ppc 970 or newer, while little endian targets
are set to a generic powerpc64le which effectively means POWER8 and
newer. Tuning is always set for POWER9, which is the most likely target
hardware. We also make sure AltiVec is always on, because it is supported
on all hardware we target.

https://github.com/void-linux/void-packages/commit/151f876627ad56a735615e27c4473c18e42b2a9a