grub: enable ppc64 builds (without a native platform for now)
The correct way would be to build grub-ieee1275 but that needs a
compiler capable of generating 32-bit big endian binaries which we
don't have. It is, however, still useful to build grub utils as they
can manage the configuration files in /boot, which can subsequently be
read by bootloaders such as petitboot (for example found on the Raptor
Talos 2 POWER9 system).

https://github.com/void-linux/void-packages/commit/a1c6f3a3b7a9b8b1a823163deef736de8b8e0d43