common/build-style/waf,waf3: explicitly pass libdir
Waf has a utility function ('lib64' in waflib/Utils.py) which either
returns an empty string or '64' depending on if either of the paths
'/usr/lib64' or '/usr/local/lib64' exist. Then, the build system itself
decides its default LIBDIR to be either /usr/lib or /usr/lib64 depending
on the result of that function, except when libdir is passed explicitly.

We don't allow lib64 in our packages. We do have the /usr/lib64 path
as that is a symlink. Therefore, do not ever allow waf to configure the
path that way.

https://github.com/void-linux/void-packages/commit/82eec0d0231ab5de83b7de45b8898963fca05c9b