rust: add support for ppc64 target
This involves a few things. One is updating the internal-llvm patch
to match what we have elsewhere, the other is updating librustc_target
code to emit ELFv2 on glibc in our case. This is not handled perfectly
(see comment) but it's best we can do for now.

Also bump, so that it gets rebuilt and cross-compiling works.

https://github.com/void-linux/void-packages/commit/b2652078b539d1e388c20060b308e987586f17ee