gcc: add patch to properly handle 64-bit long double for ppc*-musl
Musl requires gcc to be built with 64-bit long doubles as it does not
support the quad precision long double format, which is implemented in
software using two doubles.

This was already being patched, but it turns out libgcc is forcibly
being built with 128-bit long doubles because of several glibc specific
APIs it exposes. This wasn't causing any practical runtime issue, but
it was causing problems when using e.g.  the ld.gold linker to link,
because it explicitly checks the ABI tag and libgcc was tagged as 128-bit
long double.

This extra patch solves the problem by conditionally compiling libgcc
with 64-bit long doubles (only on musl) and disabling the APIs that
would have been affected (as they are for glibc only).

https://github.com/void-linux/void-packages/commit/56444230f5ed731a40b5aea09f921554f31bec59