libtool: patch the libtool script for native envs when cross
When libtool is cross compiled, it introduces assumptions about the cross
toolchain into the resulting /usr/bin/libtool script, which prevents
the script from working when used on the target machine in a "native"
environment.

The reason for this is primarily that the "short triplet" versions
of the binutils tools do not exist in the target system, only in the
crosstoolchain. Even if they did though, this would still be wrong.

This impacts some real packages, for example libtommath does not build
when built natively in an environment for which libtool was cross
compiled.

https://github.com/void-linux/void-packages/commit/653f4339274d162109af27c93a741264460f2da4