common/build-style/haskell-stack.sh: use --skip-ghc-check
This forces all haskell-stack build-style using templates to use the
system ghc and never download any binary distributions even if the
version does not match. This is usually fine as long as the stackage
used for the template is recent enough. If it's not, it should probably
be bumped anyway.

This also enables stack to work on all platforms, even those for which
stack does not offer any binary ghc downloads, as long as the system
ghc is provided, e.g. for ppc64le.

https://github.com/void-linux/void-packages/commit/f0b03f1df1bcf03d4e0503ebfe458e0c2834df2d