linux5.2: enable the option of more than 64 CPUs
With enterprise hardware now becoming cheaply available as well as the
potential upcoming Threadripper 3 CPU (depending on what AMD does with
it), it's getting very easy to surpass the current limit of 64 cpus
on void. A HP DL580 G7 off ebay will run you $500-$1000 and can be
configured with 40/80 (cores/threads). For those wanting even more power
a HP DL980 G7 will run you ~$5k, and is 80/160.

Note: This does not require the MAXSMP flag to be enabled as SMP supports
up to 512 CPUs. Also, this is only relevant to x86_64 CPUs, as other
architectures supported by Void are not capable of hitting higher values
than the default.

https://github.com/void-linux/void-packages/commit/9615d73d449834de40000b3801fd592014b2ae2e