libtool: respect environment variables
The /usr/bin/libtool script used to use cross tool-chains, when it was
cross compiled.
This broke cross-built libtool for native builds and was changed in
https://github.com/atweiden/voidpkgs/commit/190a2e7fa174f04f5df23c0d9a175695c09f8802
to native tool-chains.
Setting native tool-chains in libtools causes cross-builds which use
libtool to fail.

Thus, change /usr/bin/libtool to use environment variables when defined
and fall back to the native tool-chain otherwise.

https://github.com/void-linux/void-packages/commit/67f8177462d2b408ccac3dc5c0221b21f9495ae8