gcc: backport omp false positive fix
This fixes gcc not marking constants emitted into assembly as shared
which would result in strange build errors such as:

```
$ gcc -fopenmp test-openmp.c
test-openmp.c: In function ‘testfunc’:
test-openmp.c:10:15: error: ‘*.LC0’ not specified in enclosing ‘parallel’
   10 |         float k[3] = { 0.1f, 0.1f, 0.1f } ;
      |               ^
test-openmp.c:7:9: error: enclosing ‘parallel’
    7 | #pragma omp parallel for default(none) shared(f)
      |         ^~~
```

This happens only sometimes and not on all platforms and seems to mostly
come down to luck.

Not bumping crosstoolchains for now, as that would be too intensive on
the builders right now. If we encounter this somewhere we should bump
cross as well.

https://github.com/void-linux/void-packages/commit/7d9c212ec4d74c72b551d956ba1d43ef4a1b6fda