shutils/purge_distfiles.sh: fix purging of multiple distfiles
Previously, xbps-src would try to find checksums in template files and
not purge those found but it actually ignored everything that had the
checksums specified on multiple lines.

The reason for this is that the bash 'read' builtin reads until it finds
a newline, and if you specify multiple checksums on multiple lines rather
than on one line it would ignore all but the first.

Fix this by replacing newlines with spaces, so that it can properly read
all checksums into an array.

https://github.com/void-linux/void-packages/commit/7febf540a9ab8f49f8ff111fe30ebccdd2e0d098