cross-arm-none-eabi-gcc: update to 9.2.0 + build libgcc
This does other reworking, such as building in a separate directory like
is recommended by the project.

By importing some changes from Arch, I managed to fix building of libgcc,
which further enables building the rest of the toolchain.

However, building libstdc++ is skipped, as that requires newlib to build,
which itself is not available at this point. Newlib on the other hand
requires a built toolchain to compile. We solve this by building gcc
(complete but without libstdc++) first, followed by building newlib,
followed by building just libstdc++ on its own.

https://github.com/void-linux/void-packages/commit/541236ac9e36d54a33ae99859dc76c1bd09c9ba3