common/build-style/cmake.sh: strip -pipe if present in C(XX)FLAGS
Reason: https://gitlab.kitware.com/cmake/cmake/issues/19590

Our workaround within cmake is not sufficient as it does not address
the issue fully and things still break sometimes. So work around this
in the build-style for the time being (and drop the cmake patch).

Once this is fixed upstream (probably needs special casing for the -pipe
flag and strip it during compile tests) we can drop this.

https://github.com/void-linux/void-packages/commit/f317ed5e7ea0bdbb564d6a0e71b87b9b51c92440