inetutils: fix the missing port number for whois
Fixes issue 18445:
https://github.com/void-linux/void-packages/issues/18445

inetutils-whois consults the /etc/services provided by IANA to
determine the port number for the "whois" service. However, the
"whois" service is not listed in that file and when you run whois
you will end up with the following error:

> whois 4.2.2.4
  getaddrinfo: Servname not supported for ai_socktype

This patch makes use of "nicname" port which is the same number
(43) as a "whois" port and is listed in /etc/services.
Moreover, changing the whois/whois.c leads to invocation of
help2man. This is unnecessary, because no option has changed and
launching help2man during the cross compilation fails.

This fix wouldn't be possible without the help of great folks
at #voidlinux.

https://github.com/void-linux/void-packages/commit/3e854219e5b7c489bf9bac87a1f54512cbeca57f