xbps-src: prevent builddir/destdir from always being removed
The problem here is that newly introduced behavior in commit
9f1ea06d16b68d9bc72a7f17d68ac33246146cad resulted in builddir and
destdir *always* being removed when starting a build, as a part of
masterdir auto-update. We don't want this as we may want to resume
a previous build (e.g. by running stages individually or by resuming
a failed build from where it stopped).

Therefore, explicitly override the removal to restore previous behavior.

https://github.com/void-linux/void-packages/commit/4098c62ebcb2fcee942f490d77ea276d673840e1