bsdtar, tar: remove tar alternative
Alternative for tar was introduced to make switch from gnu tar to bsdtar
in base-devel easier.

However, bsdtar has different interface than gnu tar, and with
alternatives using gnu tar as tar is actually harder when it is needed.

There are also many user-used scripts assuming gnu tar, which are broken
now. They should not be affected because of internal matters of packaging.

https://github.com/void-linux/void-packages/commit/9816b3d831a9e3ed744865fdd139a138a7ce904e