environment/build-style/go.sh: specify a default set of archs
This is necessary so that crossbuilds to unsupported architectures are
not actually attempted. There is a default archs set which covers all
architectures supported by the official compiler, and setting it in the
template can be used to restrict it more.

Also, add missing GOARCHs. These are irrelevant to whether we can
currently build for that arch or not. Just keep it around as a list
of potential archs to support. These are taken from gccgo, and in case
support for anything is added in the official compiler, they should match.

https://github.com/void-linux/void-packages/commit/ec937e4d293c642dc45c20deef479dc79d80751a