ghc: update to 8.8.3
Incorporates https://github.com/void-linux/void-packages/pull/20368

Also uses a recent libffi tarball for generating bindists.
This may not matter so much for x86_64 but the libffi tarball
ghc bundles by default is from 2017 and there is an assortment
of bugs that would be too much effort to patch. Instead just fetch
a fresh, confirmed-to-work tarball from upstream source.

Currently selected revision is 4d6d286 because that's almost just
the official 3.3 release, but with a bunch of fixes applied (which
we are patching in in our own libffi). It was selected to predate
some more invasive changes made in libffi master.

https://github.com/void-linux/void-packages/commit/bbbb1efebc35654b81368256ba429232d514ffe8