xbps-src: fix show_pkg_build_depends after af6e1a7.
7fed360 introduces dependencies based in subpkgs,
but this makes "xbps-src show-build-deps" return cyclic
dependencies in some cases, i.e:

$ ./xbps-src show-build-deps kirigami2
cmake
extra-cmake-modules
kcoreaddons
kirigami2
qt5
$

This is because `kirigami2-devel` depends on `kirigami2`, thus
introducing a cycle.

Skip cycles completely, note that "show-build-deps" should not
be used to collect deps for cross compilation, because you can't
know what type of dependency it is: host or target.

https://github.com/void-linux/void-packages/commit/a73783d17da9552c44702cab6fe5fb16c8ac2b40