xbps-src: add numpy build_helper
Packages that link against python3-numpy require special consideration
to ensure that, in a cross-building environment, target libraries and
headers as well as the FORTRAN compiler and linker are properly
identifeid. Previously, this was done directly in the python3-scipy
template to support cross compilation. Now, a build_helper generalizes
these changes to support other package templates.

https://github.com/void-linux/void-packages/commit/8e6c1786d1c493d4094e96181454a2a0e35c4485