chroot.sh: export XBPS_ALT_REPOSITORY to chroot
Outside of chroot, we calcucate XBPS_ALT_REPOSITORY only to throw it
away and calculate it again inside chroot. Normally, this is not problem
except when we're working with a Git's worktree, in which case, we don't
have the access to original gitdir inside chroot.

Let's bring that value from outside of chroot into it.

https://github.com/void-linux/void-packages/commit/08a382be7e08034d1baea19f31920726430058c5