linux-firmware: broadcom
linux-firmware: move Raspberry Pi firmware files to rpi-firmware-network

Now that linux-firmware-network provides the necessary wifi blobs for
the Raspberry Pi, we can remove them from rpi-firmware and provided them
in a new rpi-firmware-network subpackage.

The linux-firmware-network subpackage is modified so that only the brcm
not related to the Raspberry Pi is included.

https://github.com/void-linux/void-packages/commit/2a2c456e5e20c5bb63698cefc3d0bd9720631a73

linux-firmware: subpackage for broadcom, and remove weird rpi one

also properly install the license

https://github.com/void-linux/void-packages/commit/6fdd2b9f37d1abe9bc3973a2a44931230c57fa97