bind: fix named service.
When used with the -f option, it can crash, likely due to a lack of
syslog. Using -g makes it work but still keeps it in the foreground, by
logging directly to stderr. By adding a logger to it, we follow the
pattern of most other services.

Fixes https://github.com/void-linux/void-packages/issues/18961.

https://github.com/void-linux/void-packages/commit/6041f96bef58174255e6db66428a3c866e59dab6