gcc: don't throw exception on locales other than "C" on musl.
libtsdc++'s generic locale implementation is stuck in ~2003,
unfortunately, and just throws an exception if locales other than "C"
are used. Some applications catch this error, but they shouldn't have
to, so we will import a patch to disable the exception. It is a
workaround, and the best solution is adding a proper generic
implementation to libstdc++.

Fixes https://github.com/void-linux/void-packages/issues/18659

https://github.com/void-linux/void-packages/commit/d4a9febf12a5d5cdfbee54703335e8dd1fe98848