common/build-style: remove go_get from Go style.
Fix Manual accordingly, and also fix indentation to be compatible with
nearby items.

The two packages which set this variable set it explicitly to "no", so
it wasn't relied upon. From its description, it was recommended only for
git packages, which by default don't fit Void's packaging guidelines.
Removing to avoid anyone coming to rely on it.

https://github.com/void-linux/void-packages/commit/d9c83dbe8be08bd7ffbebc18887a6300b9b3a3b6