lua52: -devel only depends on self.
Only the lua binary depends on libreadline and ncurses, not the library,
so lua52-devel doesn't need to depend on their devel packages.

While we are here, remove ncurses-devel from makedepends, since it's no
longer used by the lua build system (this doesn't make a huge
difference, because readline-devel still pulls it in).

https://github.com/void-linux/void-packages/commit/8baff5dbd69fb524f87c8b4161fc528af865f383