lua54: -devel only depends on self.
Only the lua binary depends on libreadline and ncurses, not the library,
so lua54-devel doesn't need to depend on their devel packages.

While we are here, remove ncurses-devel from makedepends, since it's no
longer used by the lua build system (this doesn't make a huge
difference, because readline-devel still pulls it in).

https://github.com/void-linux/void-packages/commit/d7bb0801c6790471bf1a9d90815546831d718d96