p11-kit: fix trust paths.
Some flatpaks broke with the move to meson [1], which doesn't set
trust_paths automatically (the autoconf based build system did, by
checking what files were available on the system - it picked
/etc/ssl/certs/ca-certificates.crt for our build).

p11-kit on the host system matters for flatpaks, as seen in [2].

Thanks to steinex for coming up with the solution. [3] was a fix for the
same issue, but with autoconf.

[1] 0602980b08967a5f1d1b94b7d08d7d0925420ddb
[2] https://github.com/p11-glue/p11-kit/issues/275
[3] fff19573410c8d61c88b960cd117e92b63729bf1

https://github.com/void-linux/void-packages/commit/943e49052ccaefccddf9005b8c4048b81f2bdc58