libmicrohttpd: update to 0.9.73.
This update breaks some C++ implementations (see [1] and
https://github.com/void-linux/void-packages/commit/05ed2cb76ba90544e9910e5dbbcabd19d97a1a99).

All packages that depend on libmicrohttpd have been checked; all but two
(psensor and retroshare) build against libmicrohttpd >= 0.9.71 by now,
and those two are easy enough to patch.

The relevant change is just switching some API return types from `int`
to an enum with the same values, so there isn't any concern about ABI
compatibility with the change (and in fact it's only a warning on a C
compiler).

[1] https://github.com/kismetwireless/kismet/issues/281

https://github.com/void-linux/void-packages/commit/a6f4c6166b88c0777b941e69f0378fdead65b648