zlib: remove LDFLAGS workaround.
Added in https://github.com/void-linux/void-packages/commit/4d39b701666370b53a7167c00702a1d7a4213a87 without explanation as
to why.

We also revbump to rebuild the binary. On x86_64-musl, old libz.so,
built with gcc 7, is using a 2132KB RE mmap when dynamically linked,
probably because of a 0x200000 alignment required for the RE section.
After rebuilding, 96KB are used, with the actual RE mmap reduced to 56K.
The new alignment is 0x1000.

https://github.com/void-linux/void-packages/commit/9dc9a42eb7c33bbfa4e3cec21a743d4bc05114b4