musl-bootstrap: only build for glibc.
Some of this package's files conflict with the normal musl package,
which would break a system if it's installed and then uninstalled,
ending up without the dynamic linker.

Furthermore, it's only useful for glibc, anyway. musl archs already have
the normal musl package.

Since we are here, lint package and remove outdated sed for when cc was
still an alternative.

https://github.com/void-linux/void-packages/commit/e1c153d8416da00ff960e635f54a1742fa33fdb3