screen: fix launch.
When screen isn't suid, it can't create the directory under /run/screens
for sessions to sit in. We went with the approach of creating the
directory at boot time instead of removing --with-socket-dir so screen
would default to using ~/.screen, because it would require users with
currently running screen sessions to launch the screen client with
SCREENDIR set correctly, and an INSTALL.msg to warn users.

https://github.com/void-linux/void-packages/commit/56db43e8323f014d5c7791bbf965fba0dcdb742e