util-linux-common: update to 2.37.1.
Rename the ${sourcepkg} to util-linux-common to avoid the trickiness
of do_configure. When util-linux-libs was invented, I was puzzled
about picking something to be the ${sourcepkg} but none of them
can satisfy the requirements.

Turn out, some of util-linux's libraries use those translated messages.

Let's invent a new package: util-linux-common, and make it the sourcepkg
for those libraries.

This package will contain all the translated message for util-linux.

https://github.com/void-linux/void-packages/commit/1bb5e87f3fac3b64f94c4af4c36b40dae4d86a8d