chroot: create chroot-curl subpackage.
The new potential build infrastructure needs a curl binary to send build
results to outside the chroot, but we don't want it to be picked up by
build systems. libcurl is part of the masterdir environment already,
through binutils -> libdebuginfod -> libcurl, so we can simply copy the
binary with a different name, and have the package also depend on
libcurl.

https://github.com/void-linux/void-packages/commit/5ab63bbf68e98aa24b44347ab43760e2ccd19519