build-styles: add zig-build
We call this "zig-build" instead of just "zig" as this build-style
relies on usage of the zig build system. In the future, other build
systems such as meson may support zig code. Furthermore, the zig
build system may be used to build C/C++ code as well, not just zig.

https://github.com/void-linux/void-packages/commit/020e1aa54bec5f077b47dda355eed837c24467e6